UPLOAD INTEGRATION INTO INTERFACE HAS BEGAN
The greatest adventure is what lies ahead.
Today and tomorrow are yet to be said.
The chances, the changes are all yours to make.
The mold of your life is in your hands to break.

The greatest adventure is there if you're bold.
Let go of the moment that life makes you hold.
To measure the meaning can make you delay;
It's time you stop thinkin' and wasting the day.

The man who's a dreamer and never takes leave
Who thinks of a world that is just make-believe
Will never know passion, will never know pain.
Who sits by the window will one day see rain.

The greatest adventure is what lies ahead.
Today and tomorrow are yet to be said.
The chances, the changes are all yours to make.
The mold of your life is in your hands to break.

The greatest adventure is what lies ahead...




git-svn-id: https://xp-dev.com/svn/crackOS/branches/production@369 2def0f6a-7ed5-4680-9172-106844e829bd